Backend/requirements.txt
Django==5.2.11
djangorestframework
django-cors-headers
pdfplumber
spacy